+++ spec/models/notifier_spec.rb
@@ -10,23 +10,37 @@ describe Notifier do
 
   describe "#send_message" do
     let(:message) { "hello world" }
-    let(:options) { {} }
-
-    it "passes the message along with some default options" do
-      expect(client).to receive(:chat_postMessage)
-        .with(text: message,
-              username: "Reminders App",
-              icon_emoji: ":loudspeaker:")
-      notifier.send_message message, options
+    let(:default_message) do
+      { text: message, username: "Reminders App", icon_emoji: ":loudspeaker:" }
     end
 
-    it "passes additional options to client" do
-      options = { foo: :bar }
-      expect(client).to receive(:chat_postMessage)
-        .with(text: message,
-              username: "Reminders App",
-              foo: :bar, icon_emoji: ":loudspeaker:")
-      notifier.send_message message, options
+    context "when options contain no channel" do
+      let(:options) { {} }
+
+      it "doesn't send any messages to client" do
+        expect(client).not_to receive(:any)
+        notifier.send_message message, options
+      end
+    end
+
+    context "when options contain one channel" do
+      let(:options) { { channel: "chan" } }
+
+      it "sends one message to client" do
+        expect(client).to receive(:chat_postMessage).with(default_message.merge(channel: "#chan"))
+        notifier.send_message message, options
+      end
+    end
+
+    context "when options contain 3 channels" do
+      let(:options) { { channel: "chan1 chan2 chan3" } }
+
+      it "sends 3 messages to client" do
+        expect(client).to receive(:chat_postMessage).with(default_message.merge(channel: "#chan1"))
+        expect(client).to receive(:chat_postMessage).with(default_message.merge(channel: "#chan2"))
+        expect(client).to receive(:chat_postMessage).with(default_message.merge(channel: "#chan3"))
+        notifier.send_message message, options
+      end
     end
   end
 end